+++ src/ColumnsContainer.php
@@ -16,7 +16,7 @@ namespace MarcinOrlowski\AsciiTable;
 
 use Traversable;
 
-class Columns implements \Countable, \ArrayAccess, \IteratorAggregate
+class ColumnsContainer implements \Countable, \ArrayAccess, \IteratorAggregate, ArrayableContract
 {
     /** @var Column[] $columns */
     protected array $columns = [];
@@ -41,6 +41,9 @@ class Columns implements \Countable, \ArrayAccess, \IteratorAggregate
         }
 
         $this->columns[ $columnIdx ] = $column;
+
+        $this->get($columnIdx)->updateMaxWidth(\strlen($column->getTitle()));
+
         return $this;
     }
 
@@ -93,5 +96,12 @@ class Columns implements \Countable, \ArrayAccess, \IteratorAggregate
     public function getIterator(): Traversable
     {
         return new \ArrayIterator($this->columns);
+    }
+
+    /* ****************************************************************************************** */
+
+    public function toArray(): array
+    {
+        return $this->columns;
     }
 }