+++ src/Renderer.php
@@ -14,11 +14,13 @@ declare(strict_types=1);
 
 namespace MarcinOrlowski\AsciiTable;
 
+use MarcinOrlowski\AsciiTable\Output\OutputContract;
+
 class Renderer
 {
     /* ****************************************************************************************** */
 
-    public function render(Table $table, OutputContract $writer): void
+    public function render(AsciiTable $table, OutputContract $writer): void
     {
         $columns = $table->getColumns();
 
@@ -36,21 +38,30 @@ class Renderer
                 $writer->write($this->renderRow($columns, $row));
             }
         } else {
-            $writer->write('NO DATA');
+            $label = 'NO DATA';
+            if (\strlen($label) > $table->getTotalWidth()) {
+                $label = \substr($label, 0, $table->getTotalWidth() - 1) . '…';
+            } else {
+                $label = \str_pad($label, $table->getTotalWidth(), ' ', \STR_PAD_BOTH);
+
+            }
+            $noData = \sprintf('| %s |', $label);
+            $writer->write($noData);
         }
         $writer->write($sep);
     }
 
     /* ****************************************************************************************** */
 
-    protected function renderRow(Columns $columns, Row $row): string
+    protected function renderRow(ColumnsContainer $columns, Row $row): string
     {
         $result = '';
-
         $cells = $row->getCells();
         $cnt = \count($cells);
         $columnOffset = 0;
-        foreach ($cells as $columnKey => $cell) {
+        foreach (\array_keys($columns->toArray()) as $columnKey) {
+            $cell = $cells->get($columnKey);
+
             /**
              * @var string|int $columnKey
              * @var Cell       $cell
@@ -68,6 +79,8 @@ class Renderer
             $columnOffset++;
         }
 
+        $result .= PHP_EOL;
+
         return $result;
     }
 
@@ -77,7 +90,7 @@ class Renderer
     protected const HEADER_PAD_CENTER = ' | ';
     protected const HEADER_PAD_RIGHT  = ' |';
 
-    protected function renderHeader(Columns $columns): string
+    protected function renderHeader(ColumnsContainer $columns): string
     {
         $result = '';
 
@@ -86,13 +99,14 @@ class Renderer
         foreach ($columns as $columnKey => $column) {
             /**
              * @var string|int $columnKey
-             * @var Column       $column
+             * @var Column     $column
              */
             if ($columnOffset === 0) {
                 $result .= self::HEADER_PAD_LEFT;
             }
 
-            $result .= $this->pad($columns, $columnKey, $column->getTitle());
+            $align = $this->getColumnTitleAlign($columns, $columnKey);
+            $result .= $this->pad($columns, $columnKey, $column->getTitle(), $align);
 
             $result .= ($columnOffset === $cnt - 1)
                 ? self::HEADER_PAD_RIGHT
@@ -100,6 +114,8 @@ class Renderer
 
             $columnOffset++;
         }
+
+        $result .= PHP_EOL;
 
         return $result;
     }
@@ -110,7 +126,7 @@ class Renderer
     protected const HEADER_SEGMENT_CENTER = '-+-';
     protected const HEADER_SEGMENT_RIGHT  = '-+';
 
-    protected function renderSeparator(Columns $columns): string
+    protected function renderSeparator(ColumnsContainer $columns): string
     {
         $result = '';
         $cnt = \count($columns);
@@ -133,46 +149,56 @@ class Renderer
             $columnOffset++;
         }
 
+        $result .= PHP_EOL;
+
         return $result;
     }
 
     /* ****************************************************************************************** */
 
-    protected function pad(Columns $columns, string|int $columnKey, string $value): string
+    protected function pad(ColumnsContainer $columns,
+                           string|int       $columnKey,
+                           string           $value,
+                           ?Align           $align = null): string
     {
-        $width = $this->getColumnWidth($columns, $columnKey, $value);
-        $align = $this->getColumnAlign($columns, $columnKey);
-        return $this->padRaw($value, $width, $align);
-    }
+        // If no custom align specified, inherit column's default align.
+        $align ??= $this->getColumnAlign($columns, $columnKey);
+        $maxWidth = $this->getColumnWidth($columns, $columnKey, $value);
 
-    protected function padRaw(string $string,
-                              int    $minWidth,
-                              Align  $align = Align::RIGHT,
-                              string $padding = ' '): string
-    {
         $padType = match ($align) {
-            Align::LEFT => \STR_PAD_RIGHT,
             Align::RIGHT => \STR_PAD_LEFT,
+            Align::LEFT => \STR_PAD_RIGHT,
             Align::CENTER => \STR_PAD_BOTH,
             Align::AUTO => \STR_PAD_RIGHT,
         };
 
-        return \str_pad($string, \max(\strlen($string), $minWidth), $padding, $padType);
+        $strLen = \strlen($value);
+
+        // Clip the string if it is longer that max allowed column width
+        if ($strLen > $maxWidth) {
+            $value = \substr($value, 0, $maxWidth - 1) . '…';
+        }
+
+        return \str_pad($value, $maxWidth, ' ', $padType);
     }
 
     /* ****************************************************************************************** */
 
-    protected function getColumnWidth(Columns $columns, string|int $columnIdx, string $value): int
+    protected function getColumnWidth(ColumnsContainer $columns, string|int $columnIdx, string $value): int
     {
         $columnMeta = $columns->get($columnIdx);
-        return \max($columnMeta->getWidth(), \strlen($value));
+        return $columnMeta->getWidth();
     }
 
-    protected function getColumnAlign(Columns $columns, string|int $columnIdx): Align
+    protected function getColumnAlign(ColumnsContainer $columns, string|int $columnIdx): Align
     {
         return $columns->get($columnIdx)->getAlign();
     }
 
+    protected function getColumnTitleAlign(ColumnsContainer $columns, string|int $columnIdx): Align
+    {
+        return $columns->get($columnIdx)->getTitleAlign();
+    }
     /* ****************************************************************************************** */
 
 }