+++ CHANGES.md
@@ -6,9 +6,18 @@ Fast and flexible PHP library rendering plain ASCII tables.
 
 ## Changes
 
-* v.1.1.0 (2022-10-17)
-  * Corrected `\ArrayAccess` implementation
+* v1.2.0 (2022-10-17)
+  * Main class name changed to `AsciiTable`.
+  * Added option to configure column default align.
+  * Added option to separately control table column title align.
+  * Custom column max width is now correctly handled.
+  * Fixed width columns are now truncated automatically with ellipsis.
+  * Improved code creating row cells from plain string arrays.
   * Added more tests.
 
+* v1.1.0 (2022-10-17)
+  * Corrected `\ArrayAccess` implementation.
+  * Added more tests.
+  
 * v1.0.0 (2022-10-17)
   * Initial public release.